+++ dev/tests/performance/framework/tests/unit/testsuite/Magento/Test/Performance/BootstrapTest.php
@@ -19,7 +19,9 @@ class BootstrapTest extends \PHPUnit_Framework_TestCase
         // Delete a directory, where tests do some temporary work
         $tmpDir = $this->_getBaseFixtureDir() . '/config_dist/tmp';
         $filesystemAdapter = new \Magento\Filesystem\Driver\File();
-        $filesystemAdapter->deleteDirectory($tmpDir);
+        if ($filesystemAdapter->isExists($tmpDir)) {
+            $filesystemAdapter->deleteDirectory($tmpDir);
+        }
     }
 
     /**
